PardotApi-v1-NewDashboard/Backend/requirements.txt
flask
flask-cors
requests
reportlab
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dateutil
PyJWT
redis